requirements.txt
# Core dependencies
streamlit>=1.28.0
pandas>=2.0.0
plotly>=5.15.0
numpy>=1.24.0

# Machine Learning
torch>=2.0.0
transformers>=4.30.0
datasets>=2.14.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0